Move common nav item to dedicated component
